Split cms.handlers into per-domain handler modules (#1068)
Replaces shifter/shifter_platform/cms/handlers.py (389 LOC, mixing five
responsibilities) with a cms/handlers/ package: envelope (SNS unwrap),
range_events, experiment_bridge, ctf_bridge (signal fire), ngfw_events,
plus a dispatcher in __init__. Public surface preserved via re-exports —
cms.handlers.process_event (the SQS worker entry referenced as a string
in config/settings.py), parse_sns_message, process_range_event,
process_ngfw_event, and notify_experiment_on_range_ready all keep their
existing import paths.

Behavior unchanged: prefix routing, request_id-over-range_id correlation,
range_id non-overwrite, log-and-return on bad data, signal wiring through
cms.signals.range_status_changed, and FAILED-on-publish-error semantics
for the experiment bridge are all byte-equivalent.

Existing test patch targets moved to point at the new owning submodules
(cms.handlers.range_events.X, cms.handlers.ngfw_events.X,
cms.handlers.experiment_bridge.X) — no test logic changed. Adds a
TestProcessEvent case for the previously-untested experiment.* route.

Sub-issue of #1055.
